Add option to fail gracefull on Rabit::Init failure
When running xgboost4j-spark the process that rabit
is initialized in has other things going on, such as
caches of data that is expensive to recompute. On
the rare occasion that rabit fails to connect to the
tracker rabit performs an exit(-1) which throws away
everything that was going on in the application.

It would be nice to somehow provide explicit failure
handling everywhere, but changing the external api
to that level would be quite a large breaking change.
This patch takes a very targeted approach changing
only the Init call to return a boolean indicating
success. This is disabled by default and must be
provided as part of the initialization parameters.